#35: Consolidate defaults and constants for SKAWorkflows.
There are values that we get from the Parameteric model or other places that are very useful for grounding us to previous research. We have a lot of these defaults around the place and we shouldn't be copying them; instead, we want a single source of truth.

This sets up those single sources of truth, as well as (hopefully) starting to codify the user-interfaces with things like the default observing parameters.

I have also removed the build directory from what is committed. Hopefully this is no longer in future commits.

Signed-off-by: myxie <contact@ryanbunney.com>